+++ Changelog.md
@@ -1,8 +1,21 @@
-# v0.2.0 (2013-31-03)
 
-#### Breaking Changes:
+# v0.2.1 (2014-06-02)
 
-- API change: create instances of EditableJS [65](https://github.com/upfrontIO/Editable.JS/pull/65)
+#### Features
+
+- Add change event [#66](https://github.com/upfrontIO/Editable.JS/pull/66)
+- Force height of empty elements (especially in Firefox) [#68](https://github.com/upfrontIO/Editable.JS/pull/68)
+
+#### Bugfixes
+
+- Set Focus in iFrame properly [657f85](https://github.com/upfrontIO/Editable.JS/commit/657f85d1c1a0f9d3018548654271616c41480b2b)
+
+
+# v0.2.0 (2014-03-31)
+
+#### Breaking Changes
+
+- API change: create instances of EditableJS [#65](https://github.com/upfrontIO/Editable.JS/pull/65)
 
 
 # v0.1.2 (2014-03-07)